chore(dropdown): align styling to multi select (#57)
This change aligns the markup/styling of dropdown to multi select. In
this way, the dropdown can share many portions of its code with multi
select.

Given this change transitions the markup and the interaction model from
one from Carbon `v6` to one from `carbon-components-react` `v10`, the
new `dropdown.ts` code is closer to multi select than the existing
`dropdown.ts` code.

The new dropdown code keeps focus on the trigger button even while the
menu body is open, and takes a manual approach for screen reader, which
is, putting content to announce into a special element that lets screen
reader read it whenever a new content is put in.