chore(build): remove bundled build (#111)
This change removes the bundled build. Though bundled build is easier
to use, it requires external dependencies to be either bundled in or
defined in global scope, where we want to avoid doing the former
wherever it makes sense and the latter is not the case with any
`lit-element` OOTB deliverables.

We can re-consider bundled build when
[import maps spec](https://github.com/WICG/import-maps) or something
similar is available in major browsers.