test(components): add snapshot tests (#126)
This change introduces a Jasmine custom matcher, `toMatchSnapshot()`, to see if the given element or the shadow DOM content of it (with `toMatchSnapshot({ mode: 'shadow' })`) matches its snapshot.

The snapshot is managed by `karma-snapshot`. The snapshot HTML is normalized by `@open-wc/semantic-dom-diff`. `jest-diff` formats the diff if the snapshot test fails.

This change also introduces a test against `<bx-checkbox>` using such snapshot feature.