chore(dropdown): re-introduce invalid attribute (#128)
In the discussion of `<bx-input>` introduction PR (#87), we decided
that we are keeping both `invalid` and `validity-message`, at least
for now, given there is not much connection in the meaning of those
attributes.

This change also introduces support for `<span slot="label-text">` and
`<span slot="helper-text">` for richer label/helper text support.