feat(form): support formdata event (#154)
This change implements `formdata` event handlers (see PR# 4239 in
whatwg/html repo) in components that should participate in form
submission.

This change allows `<form>` to gather form data in our components with
`.submit()`, `<button type="submit">`, etc. with browsers supporting
such spec, or by firing a custom event with a similar interface, while
the underlying `<input>` in our components are in shadow DOM and that
`<form>` is not normally aware of.