fix(code-snippet): fix scroll overflow UI (#222)
This change fixes the overflow UI of code snippet by fixing missing
`::after` pseudo element on shadow hosts, along with a fix for Safari
that treats `transparent` as `rgba(0, 0, 0, 0)`.

This change also ensures that the production build uses the PostCSS
plugin to fix `:hover`, etc. pseudo classes on shadow hosts.