chore(components): avoid reference to stale async results
This change uses `useEffect()`'s "clean-up" API to detect if another
`useEffect()` call will come or the component is being unmounted, to
see if a return value from an asynchronous call is stale.

This ensures only one asynchronous call is in-flight, and avoids
React's `Can't perform a React state update on an unmounted component.
This is a no-op, but it indicates a memory leak in your application.
To fix, cancel all subscriptions and asynchronous tasks in a useEffect
cleanup function.` error.