docs(react): add CodeSandbox example
This change introduces a CodeSandbox example for masthead.

Using CodeSandbox's Node.js template (instead of `create-react-app`
template) for better stability and better flexibility.

The style is broken because `styles.scss` in this example imports a
wrong Sass file, but changing it to the right file causes a build error
described in #2436.

Refs #2198.