docs(components): introduce Storybook `addon-docs` (#2480)
This change adds Storybook `addon-docs` to our React storybook
environment, starting with `<CTA>` component.

`addon-docs` harvests React DocGen data from our components and allows
us to put in our doc. To use that feature, this change converts the
`README` to `.mdx` and replaces the properties table in the `README`
with the auto-generated one (`<Props of={CTA} />`).

This change also disables `jsdoc/require-jsdoc` ESLint rule for our
React codebase, and introduces a custom ESLint rule to check for React
prop types comment. Given we haven't completed the transition, the new
ESLint rule is set as a warning (instead of an error) for now.

Refs #2422.