feat(beta-releases): adding beta release creation script (#2599)
* chore(percy): hiding video link in linklist for percy tests

We have been getting a lot of false positives with the video link in
LinkList, so going to add this to suppress that item entirely instead.

* feat(beta-releases): adding beta release creation script

This is a script that will find the diff between the previous tag and
latest, then will generate a new release on the new tag. Right now,
it will only add features and fixes and sort accordingly.