fix(storybook): bug happening in the masthead iframe view (#2593)
The problem here is that we have a bug within the storybook iframe happening with some of our stories when we change a prop in the knobs panel and then open the component in the iframe. Since we cannot fix it in our end, a workaround is to create a new story with the properties we want to see.