fix(corsproxy): adding REACT_APP_CORS_PROXY as an option for services
This will include REACT_APP_CORS_PROXY as an option for adopters that
are using `create-react-app`. Because CRA only allows environment
variables that begin with `REACT_APP_`, the only way around this would
be to eject then customize the `env.js` configuration to allow
`CORS_PROXY`. This will hopefully make it easier to adopt the
`DotcomShell` in an OOTB CRA application.