fix(corsproxy): removing cors proxy on production environments (#2633)
This will make the proxy server blank if the call is being made from www.ibm.com

In addition, this will include REACT_APP_CORS_PROXY as an option for adopters that
are using create-react-app. Because CRA only allows environment variables that
begin with REACT_APP_, the only way around this would be to eject then customize the env.js configuration to allow CORS_PROXY. This will hopefully make it easier
to adopt the DotcomShell in an OOTB CRA application.

One thing to note, I was trying to get the unit tests to switch environment variables
in place, but the mock was not reinitializing for axios so setting this aside for now.