chore(devenv): introduce theme chooser in React
This change introduces a theme chooser Storybook add-on, as a new
package named `@carbon/storybook-addon-theme` (private for now).

The new Storybook add-on creates a Storybook tab title "Carbon Theme".
When user changes the theme there, a cross-frame messaging happens that
can be listened to with Storybook
`addons.getChannel().on('carbon-theme/current')` API.

This change also sets up the Storybook environment for our React
variant. In this setup, when such cross-frame message is received:

* The `<html>` tag gets `storybook-carbon-theme` attribute with the
  theme name as its value. It works with new Sass code for Storybook
  preview pane that defines CSS custom properties for each theme,
  switched by such attribute in `<html>` tag. That said, CSS custom
  properties are now enabled always.
* The stories are re-rendered, so that stories can get the latest
  chosen theme from the `<html>` tag. Some of our React components take
  theme name via props and stories for those components have to
  demonstrate that props.

The theme chooser UI are disabled for some stories, notably
dotcomshell, masthead and footer.

Refs #2708.