chore(devenv): remove `.bx--visually-hidden` override
This change removes our CSS override for `.bx--visually-hidden` class
that is defined in our Storybook environment.

Setting `display:none` to that class makes the DOM element no longer
keyboard focusable that is against its purpose, which is, "hidden
visually, but navigable with screen readers".

Carbon modal's focus wrap behavior relies on elements with
`.bx--visually-hidden` keyboard-focusable to detect changes in focused
element that goes out of modal.

Refs #2725.