fix(Card): fix prop type (#2767)
`<Card>` does not support multiple types in its `cta` property. Also it
doesn't have code to support video type. This change fixes the
`propTypes` of `<Card>` to reflect that.

Also there are several components that renders `<CTA>`, but uses
specific type of `<CTA>` regardless of the `cta` data structure given
in the props. Fixed `propTypes` of those components to reflect that.

Co-authored-by: Jeffrey Chew <jeff.chew@gmail.com>