test(react): add test for bundle size
This change adds a test suite for making sure importing a module of a
React component won't cause bringing in non-necessary other part of
this library.

This is done by testing if a Rollup build with a React component module
as the entry-point yields to an empty result unless the React component
is in use.

Such Rollup build runs with:

* A (new) Babel plugin that marks sevaral contents as no side effect.
* Some module stub contents that marks their module entry point as with
  side effect, as some contents in the module may have side effects and
  thus importing the module entry point will cause unnecessary content
  included in application bundle.

Refs #2493.