fix(webcomponents): updating package to public
This is a fix to properly publish the package as a public
package in NPM.

In addition, this includes a fix for the clean task, as it is
failing the global `yarn reset` since `gulp clean` requires a
prebuilt package through gulp, otherwise will fail. This is
now using a standard `rimraf` like the other packages.

Finally, the dependencies have been moved around so that the
library and carbon dependencies are required.