fix(Card): fixed column number for small breakpoint (#2934)
The column number for Card on small breakpoint was wrong, it was occupying only two columns. I created a fix in order to fit the entire screen (4 columns), according to visual specs.

Note: The Card Link is occupying the correct column number, so I didn't touch the code relative to it.