feat(expresive-modal): introduce Web Components
Introduces `<dds-modal>`, an expressive (DDS) version of modal.
Also introduces `<dds-btn>`, an expressive (DDS) version of button, as
it's needed for demonstration.

We could use `<bx-btn>` with applying expressive tokens in application
(Storybook) code, but having a tag (custom element) that works as an
expressive version out-of-the-box will provide the full benefit of Web
Components, which is, "It just works by putting them as DOM elements".

`<dds-btn>` utilizes the expressive token proxy code to minimize the
Sass code. Did the following to do that:

* Introduce Sass maps for expressive size/type/component tokens.
* Extracted out and streamlined the code to generate CSS custom
  properties from expressive size/type/component tokens.
* A fix to the expressive token proxy code for button so it works when
  it's used within a CSS ruleset (e.g. one for
  `:host(#{$dds-prefix}-btn)`).

Refs #2988.