fix(styles): optimization fixes for gulp-sass in styles package (#3119)
The gulp-sass task was throwing out of memory errors. This is a step to mitigate the build times for the overall css artifact. To our knowledge, no adopters are using the main entry point and built artifact (nor should they), so for now this is removing sourcemaps from the build pipeline until this can get sorted out.