fix(ImageWithCaption): image is hidden in sm & md (#3063)
The image on modal was being occulted in sm and md breakpoints. I noticed a code was applying an absolute position below lg breakpoint was causing the issue. I removed the code, tested it in all resolutions (Modal and outside the modal) and the component is working well.

Note that I wasn't sure why that code was added there, but it's removal doesn't break the component since the code was only set for img element.