feat(cta): introduce text CTA for Web Components
Introduces Web Components version of text CTA.

The most notable difference from its React counterpart is how CTA links
are handled, notably with the video CTA type. `CTAMixin` automatically
sets `href` to `#` for video CTA type and fires `dds-cta-run-action`
custom event when the link is clicked with the original `href` value,
which is the video ID, in `event.detail`.

This allows a separate component like `<dds-cta-container>` to create a
media viewer dialog, instead of having to put in all media viewer logic
into text CTA UI as a dependency.