feat(Card): add link text option to clickable card (#3097)
Entire card should be clickable. The PR removes the "static" type option that existed previously which made only the link within the card clickable - so the type prop is no longer necessary. Logic that differentiates the "static" and "link" type has been removed as well.

Option to add text to the Card cta has been added.