feat(VideoPlayer): add aspect ratio overrides (#3139)
In regards to my concern about videos that are wrapped in other components such as Card, I think this is not as big a deal as I first thought. We can control this through the containing component itself and not allow the aspect ratio to be set for those videos.

This change adds several aspect ratios which can be used to override the default of 16:9.