test(e2e): adding WebdriverIO e2e test structure (#3166)
This adds the core WebDriverIO test harness, Github action, and first initial tests that were converted from the Python script versions. The strategy taken is to use direct urls for the full screen view of the canvas. This includes passing in query params if necessary, rather than adjusting knobs.

There are still remaining .py files, which need to be converted to e2e.js files and in WebdriverIO format. I think it will be ok to tackle this in separate PRs.