fix(ddo): setting the proper attribute for version in ddo (#3190)
### Related Ticket(s)

No related issues

### Description

Noticed that the version was not being set specifically on the KC test page, so this is a fix for this. I noticed that the version was being set in the wrong place, so this is to fix that issue. This would hopefully fix the issue for other applications but tested this in a local NextJS application and appears to work.

### Changelog

**Changed**

- Changing the location where the library version is set in DDO.

<!-- Deploy Previews are enabled by applying the following labels for the corresponding package: -->
<!-- *** "package: react": React -->
<!-- *** "package: web components": Web Components -->
<!-- *** "package: vanilla": Vanilla -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive, React (Expressive) -->
<!-- *** "RTL": React (RTL) -->
<!-- *** "feature flag": React (experimental) -->