test(e2e): updates to e2e tests for CI server (#3267)
### Related Ticket(s)

#3237 

### Description

This adds some fixes to the e2e tests for Jenkins. Part of this is giving Jenkins's chrome browser more time for the element to appear. The default value for `waitForExist` is 500 milliseconds, which might not be enough time for it since it is slower than a full fledge developer machine.

### Changelog

**Changed**

- Various tweaks to the e2e tests
- Setting the browser size by default so that it doesn't need to be defined in the tests