chore(automation): adding automation script for epic creation
This is an automation script that will create an epic
issue for a given component, whether "react" or
"web components", and corresponding design/prod/dev issue
templates. The motivation is that we have a large number
of upcoming component work necessary for both packages,
and this will help with lining up the work for these
going forward.