docs(masthead): made README more adopter friendly
This adds in a new ELI5/tl;dr section for adopters that want to
just grab the code and go, which will be primarily most adopters.

Sections that go deeper into the inner components are moved
under `Under the Hood`.

I also removed the preview, it felt redundant with the
Storybook canvas.