chore(services): introduce REST cache
This change introduces an in-memory cache for in-flight requests of
`LocaleAPI.getList()` and `TrasnlationAPI.getTranslation()` APIs.

It prevents multiple requests from getting in-flight, especially when
multiple request of those happens in a short amount of time. In other
words, without this change, the cache in `sessionStorage` from the 1st
call of `LocaleAPI.getList()` hasn't been populated when 2nd call of
`LocaleAPI.getList()` happens, given the response of the 1st call
hasn't come.