fix(masthead): added entry point to self-init redux store (#3579)
A first attempt to provide a way for our adopters to avoid having to
bootstrap the redux store manually. A discussion after the initial
commit in this PR determined that we ship `<dds-masthead-contaiher>`
with such OOTB store configured and create another component to
manually configure the store, that should happen in a follow-up PR.

Also, added some initial attempt for the documentation around
automatic/manual configuration for the store.