fix(side-nav): cope with large list of parent menu items
This change fixes an issue where the list of parent menu items is
larger than its overlayed child menu items and thus the bottom portion
of the parent menu items is still shown even if child menu items are on
top of them.

The problem is seen, at the point of the fix, when user navigates to
"Products and Solutions" - "Supply Chain" and then user scrolls up the
nav.

Refs #3512.