fix(MegaMenu): account for focus within the megamenu component (#3763)
### Related Ticket(s)

[MM]: When clicking on empty space, megamenu does not close causing stacking #3758

### Description

When user clicked on the empty space within the megamenu, the focus on the L0 nav item was lost causing issues with the `onBlur()` event. Also the `onClick()` event was placed on the wrapper container of the `<HeaderMenu />` which caused the menu to open and close unnecessarily. 

The `onClick()` event has been moved to specifically  the anchor tag so that the menu does not toggle its expanded state when user clicks anywhere within the menu. Adding `tabIndex` to the `<NavigationGroup />` (which is the container component for the megamenu) so that when user clicks on it, it will be captured in `event.relatedTarget` used in the `onBlur()` method otherwise it will render null. 

**BEFORE:**
![Aug-27-2020 11-22-36](https://user-images.githubusercontent.com/54281166/91462012-ae29df80-e857-11ea-89b6-09d324bacb13.gif)



**AFTER:**
![Aug-27-2020 11-13-39](https://user-images.githubusercontent.com/54281166/91461796-6c009e00-e857-11ea-95cf-6109847696b1.gif)

### Changelog

**Changed**

- add the `bx--header-menu` class to the list of elements checked in the `onBlur()` event to ensure the overlay doesn't toggle when clicking within the menu
- moved the `onClick()` function to the specific anchor tag instead of the wrapper.


<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: vanilla": Vanilla -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive, React (Expressive) -->
<!-- *** "RTL": React (RTL) -->
<!-- *** "feature flag": React (experimental) -->