docs(lang): updated documentation for setting the page language/locale
The documentation for how `LocaleAPI` is fetching the language has been
updated. This reflects the logic where the `digitalData` is the first
source of the language/locale, followed by the html lang attribute.
Because of this, it was necessary to explain what the `digitalData`
object is first, before pointing out the specific fields that the
service is looking for.

Additionally, we will be setting `IBM.com Library` as the default value
for the `version` that's being set in the `digitalData` object.