fix(masthead-menu-button): fix toggle event
This change fixes broken trigger button of masthead left nav
(`<dds-masthead-menu-button>`).

An earlier change there to watch for toggling active state of
`<dds-masthead-search>` introduced a name conflict of static property
for getting the event name for such active state, `eventToggle`, which
is used for toggling masthead left nav. Renamed them to
`eventToggleSearch` to solve the name conflict as well as to make it
explicit that it's an foreign event.

Refs #3834.