docs(contributing): reorganization of contribution content
The flow of contribution content was a bit fragmented, so this is a
first pass at giving it a better flow. The submission guidelines have
been broken out into its own document, which references both the
developer guidelines and React contribution guide. The setup and pull
request documentation also made more sense under the developer
guidelines.

There were other minor updates, which include setting up the ToCs in
places that made sense, and removed in some where the documentation was
primarily used as an include in other READMEs.