chore(github): fix to the user story template (#492)
* chore(github): fix to the user story template

The `User Story` template wasn't showing up in the issue creation page.
This might have to do with the formatting of it, so this should
hopefully fix it.

* chore(github): adding `.github` to `.prettierignore`

Prettier was reformatting the `.github` templates, which might have
caused issues with Github picking them up. We ignore this folder in our
other project, so it makes sense we add this here too.

Also updating conventional commit max lengths to match IBM.com Library
settings.

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>