feat(select): introduction (#478)
Introduces `<bx-select>` that works as a Carbon-styled `<select>`.

This works with `<bx-select-item>` and `<bx-select-item-group>`.
When they are put in `<bx-select>`, `<bx-select>` creates the
corresponding `<option>` and `<optgroup>`, respectively, into shadow
DOM. Doing so allows the shadow DOM style of `<bx-select>` to control
the style of the `<option>` and `<optgroup>`, notably the disabled
ones.

Also introduces Angular directive for `<bx-select>`.

Refs #229.