fix(footer): ensure IBM logo icon is visible
Fixes the rendering issue of IBM logo icon in footer by fixing wrong
`width` and `height` attributes.

Also fixes wrong content in the `<title>` attribute in corresponding
SVG, by avoiding to bake it in the `lit-html` SVG template for the IBM
logo icon. The approach is taken because the `<title>` content is not
the design asset, and may have to be translated in future. Due to such
reason, omitted some other attributes from the `lit-html` SVG template
for the IBM logo icon, too.

Refs #4224.