fix(footer): ensure IBM logo icon is visible (#4226)
### Related Ticket(s)

Refs #4224.

### Description

Fixes the rendering issue of IBM logo icon in footer by fixing wrong `width` and `height` attributes.

Also fixes wrong content in the `<title>` attribute in corresponding SVG, by avoiding to bake it in the `lit-html` SVG template for the IBM logo icon. The approach is taken because the `<title>` content is not the design asset, and may have to be translated in future. Due to such reason, omitted some other attributes from the `lit-html` SVG template for the IBM logo icon, too.

### Changelog

**Changed**

- Fixes to IBM logo icon:
  - Add right `width` and `height` attributes
  - Omit `<title>` and some attributes from `lit-html` SVG template