chore(services-store): added null check of services-store vendor folder
The motivation of this PR is that a `yarn reset` would fail if the
vendor folder for the service store doesn't exist, primarily during a
`gulp clean`. In order to allow the `gulp clean` to continue, it cannot
have an exit code 1 if the folder doesn't exist, and just continue with
the clean.

The full `yarn install`/`yarn build` works afterwards, as the link works
after that.