fix(megamenu): added protection to fix layout problem (#4290)
### Related Ticket(s)

fixes #4078 

### Description

What a fickle little layout problem we have here. I saw we were having trouble with this and had an idea of how to fix it. Basically, I threw in one or two wrapping divs to help protect the content, and render consistently across the board. It looks like it's working from my perspective and understanding, but let me know if I missed some consideration here.

### Changelog

**Changed**

- Added two new divs to help protect content and help fix 

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->