chore(build): add RTL bundle
Adds RTL version of pre-build bundle, along with a documentation as
well as a CodeSandbox example that demonstrates how to use LTR/RTL
versions of the pre-built bundle.

Also, some updates to existing CodeSandbox examples of pre-built
bundles, notably to use the latest version of the bundle instead of
`v0.5.0`.

Refs #4319.