chore(video-player): allow style customization for caption (#4377)
### Related Ticket(s)

Refs #4342.

### Description

This change allows components that render `<dds-video-player>`, e.g. `<dds-callout-with-media>`, to customize the style of the video caption via CSS custom properties.

It elminates the need to create an inheritance of `<dds-video-player-container>` for the styling purpose. Creating an
inheritance of `<dds-video-player-container>` involves creating one for `<dds-video-player-composite>` as well as one for `<dds-video-player>` that is costful.

### Changelog

**New**

- `--dds-video-caption-padding` and `--dds-video-caption-color` CSS custom properties.