feat(Masthead): React masthead alternate logo and tooltip options (#4401)
### Related Ticket(s)

#3906  

### Description

The masthead now accepts an optional logo options from IBM.com translation files which allows the setting of an alternate IBM logo to use, as well as an optional tooltip when the icon is hovered over. When an alternate logo is used, settings include an expiration date and allow/deny lists for controlling paths on which it can be displayed. The tooltip can be used with both default and alternate logos.

### Changelog

**New**

- `MastheadLogo` service to determine whether an alternate logo can be used based on values passed in the object
- React `Masthead` component now accepts options from IBM.com translations for an alternate logo and tooltip
- internal `ConditionalWrapper` component 

**Changed**

- add masthead logo svg dimension styles


<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->