docs(content-block): alternate ways to specify copy content
In addition to `<dds-content-*-copy>` based on `<dds-markdown>` that
automatically convers the markdown content to HTML client-side, adds a
couple of alternate ways to provide the copy content for
`<dds-content-*>` that is helpful when the target network and device
environment has limited capacity.

Also adds another guard to ensure `<dds-markdown>` won't treat
rendered-to-HTML content as the markdown source.