chore(changelog): updating changelog script to separate by folder
This is a modification of the changelog generation script to split up
the commit changes based on changes in each package folder. This will
help with separating out the logs based on the changes that went on in
the corresponding package folders in an automated way. From what I could
 tell, there is no existing open source package that can generate the
 logs in the desired way.