chore(changelog): updating changelog script to separate by folder (#4441)
### Related Ticket(s)

No related issue

### Description

This is a modification of the changelog generation script to split up
the commit changes based on changes in each package folder. This will
help with separating out the logs based on the changes that went on in
the corresponding package folders in an automated way. From what I could
 tell, there is no existing open source package that can generate the
 logs in the desired way.

The pinned issue has been updated using the script:
https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/3598

NOTE: one thing I did notice in one library during my research that I liked is where it was putting all of the line items with the same heading and grouping the changes into sub-bullets. Might look at that for an enhancement to the script, as well as alphabetically ordering the headings.

### Changelog

**Changed**

- Changelog script updates