fix(masthead): customtypeahead api fix (#5168)
### Description

Updates masthead search `customTypeAheadApi` to be fully on the application side. When using a custom API, it will no longer pass through `SearchTypeaheadAPI` at all. This allows users to pass a full API including any necessary query params that were not defined in `SearchTypeaheadAPI`. With this change the use of environment variables is no longer necessary.

### Changelog

**Changed**

- moves `customTypeAheadApi` out of `SearchTypeaheadAPI`
- update jest environment variables
- section titles no longer appear when there are no section items


<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->