feat(translate): accept diff endpoint arg for translation service (#5407)
### Related Ticket(s)

feat(masthead-container): Pass in the data endpoint as an attribute within the masthead-container Web Components #5396

### Description

Allow adopters to set the translation endpoint in the masthead as an attribute (ie. `<dds-masthead-container data-endpoint="/common/carbon-for-ibm-dotcom/translations/cloud-masthead">`)

This requires passing in an `endpoint` arg into the `loadTranslation()` action and then to the translation service methods. The session key required for caching the data will be derived from the endpoint by removing all special characters from the endpoint string.

Also changed default translation endpoint as we have copied over our translation files to a new folder

### Changelog

**New**

- set up new arg `dataEndpoint` for `loadTranslation()` action in the services-store
- set up new arg `endpoint` for `fetchTranslation` and `getTranslation` methods in the translation service

**Changed**

- change default translation endpoint

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->