fix(ContentBlockMedia): feature card cta icon src was being overwritten (#5608)
### Related Ticket(s)

[FeatureCardBlockMedium] icon does not get updated correctly #5493

### Description

The feature card CTA icon within `ContentBlockMedia` was constantly overwritten in the `_renderFeatureCard()` method. There is no need to pass in the `icon.src` within the props for `FeatureCTA` or `ContentBlockMedia` utilizing `FeatureCTA` as the icon will be determined by the `type` (local, jump, external, etc.) being passed in instead. 

NOTE: There are some components where we define which icon to use and pass the icon as part of the props (FeatureCardBlockMedium, FeatureCardBlockLarge, Card, etc).

We can now set the external, jump, download, etc icon instead of it being overwritten to the default ArrowRight20.
<img width="1482" alt="Screen Shot 2021-03-24 at 5 13 38 PM" src="https://user-images.githubusercontent.com/54281166/112386052-c0bda780-8cc6-11eb-8617-af020132f05c.png">

### Changelog

**Changed**

- make sure the `type` is passed in under the `card` object to ensure the correct icon is rendered in `_renderFeatureCard()`

**Removed**

- documentation in CTA (Feature Card CTA) and ContentBlockMedia where we pass the icon src

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->