test(e2e): fixes to table of contents and masthead e2e tests
The table of contents was not properly passing in the theme into the
component, which is causing some rendering issues.

The masthead is also noticing some flakiness in the renders regardless
of await commands for selectors. This adds in some additional time for
the overflow to render fully. In addition, the visual check between the
first and second mega menu nav items were separated out into two
separate tests as keeping them in one looked to have errors.