chore(dependencies): upgrade `sass-loader`
Upgrades `sass-loader` to the latest version that support WebPack4, so
that we can use `webpackImporter: false` option  to disable
`sass-loader`'s custom Sass importer that was one of the bottlenecks of
Storybook build.

`sass-loader`'s custom Sass importer is meant for non-standard
`@import` with `~`, what resolves dependencies in `node_modules`.
However, we are guiding our users to use `SASS_PATH` environment
variable or `includePaths` option in `node-sass`.

Also:

* Introduces `cache-loader` in there to further boost the performance.
* Upgrades `css-loader`, `node-sass` and `style-loader`.