chore(dependencies): upgrade `sass-loader` (#615)
* chore(dependencies): upgrade `sass-loader`

Upgrades `sass-loader` to the latest version that support WebPack4, so
that we can use `webpackImporter: false` option  to disable
`sass-loader`'s custom Sass importer that was one of the bottlenecks of
Storbook build.

Also introduces `cache-loader` in there to further boost the
performance.

Quick benchmark result is like:

* `webpackImporter: true`: 37s
* `webpackImporter: false`: 23s
* Using `cache-loader`: 16s

Also upgrades `css-loader`, `node-sass` and `style-loader`.

* chore(dependencies): remove `fast-sass-loader`

* chore(build): cope with option scheme change in `node-sass`

* test(styles): update Sass configuration in test