fix(same-height): using same height method for card group & cta section (#5567)
### Related Ticket(s)
#5398

### Description
The current CSS grid implementation for both CardGroup and CTA Section work are able to resize properly dynamically depending on the copy content, however it does not work if the heading is longer than a single line. After some discussion, a CSS-only way might actually not be achievable, thus this PR addresses that by bringing back the `sameHeight` utility function for both `CardGroup` and `CTASection`, only targeting the item headings. 

### Changelog

**Changed**

- re-applying the `sameHeight` utility function for `CardGroup` and `CTASection` targeting item headings

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->