fix(table-of-contents): fixed active section (#5794)
### Related Ticket(s)
#5498

### Description
The current Table of Contents changes the active section on the left the moment a new anchor tag appears in the viewport, even if it hasn't reached the top yet. 

| React | Web components |
| --- | --- |
| ![react-scroll-selected-section-Mar-15-2021 15-04-29](https://user-images.githubusercontent.com/45692622/111207681-4db68180-85a0-11eb-86b9-6a7edb1ff6cf.gif)  |  ![webcomponent-scroll-selected-section-Mar-15-2021 15-05-51](https://user-images.githubusercontent.com/45692622/111207628-3f686580-85a0-11eb-8875-cf713d42638d.gif) |

This PR changes this behavior so the current section is only set to "active" upon it reaches the top, and if a user scrolls down a bit further, the previous section will be the one becoming active, as shown in React.

The Web Component version of ToC uses Intersection Observer API for this functionality. However, after various attempts to modify the existing observer to apply this effect properly, it seems that there are times when the Intersection Observer API fails to detect the element coming into the viewport.

For example, one can be scrolling through Section 1, pass through Section 2 without 2 becoming active, and upon reaching Section 3 this is the one that becomes active, i.e. it ignores Section 2 completely. This happens randomly, and due to the spontaneous nature of this bug, it might be better to stick with using a Scroll Listener, like React is already doing.

### Changelog

**Changed**

- replaced Intersection Observer with Scroll Listener to align with React behavior

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->