fix(locale): return locale with the fetch list call (#5899)
### Related Ticket(s)

Masthead and Footer components not supporting locales like en-cn or en-gb #5890

### Description
For pages with DDO set to a locale that we do not support (ie. cn-en), the masthead and footer is returning blank.

When fetching the locale from the DDO to use in the translation service, we make a call to fetch the countrylist with the retrieved locale. If the countrylist call fails, we make another call with our default locale. However this change to the default locale is not communicated to the translation service and so the translation call is made with the non-existent locale.

Solution is to set the locale in the data returned from the successful countrylist call and use that locale.

### Changelog

**Changed**

- set locale in the returned countrylist data in `fetchList()`
- use the locale from the returned data

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->