fix(locale): use getLocale method to verify locale before passing into translation call (#5907)
### Related Ticket(s)

Masthead and Footer components not supporting locales like en-cn or en-gb #5890

### Description

In web-components masthead, the masthead returns blank when using a non-existent locale (ie. cn-en).

For web-components, we are fetching the locale using `LocaleAPI.getLang()` which doesn't verify the locale and passing the unverified locale directly to the translation call as a parameter. When passing the locale as a parameter, there are no checks are the translation call is made with the unverified locale. 

Solution is to use `LocaleAPI.getLocale()` instead which verifies the locale before returning it.

### Changelog

**Changed**

- For `loadLanguage()` redux action in services-store, use `getLocale()` instead to make sure the locale retrieved from DDO is valid.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->